Crate deps_lsp

Re-exports

pub use server::Backend;

Modules

config

document

Document management module.

file_watcher

Lock file watching infrastructure.

handlers

LSP protocol handlers.

progress

LSP Work Done Progress protocol support for loading indicators.

server

Structs

ArtifactInfo

BundlerDependency

Parsed dependency from Gemfile with position tracking.

BundlerEcosystem

Bundler ecosystem implementation.

BundlerParseResult

Result of parsing a Gemfile.

BundlerVersion

Version information for a gem from rubygems.org.

CargoEcosystem

Cargo ecosystem implementation.

CargoParser

Parser for Cargo.toml manifests implementing the deps-core traits.

CargoVersion

Version information for a crate from crates.io.

CrateInfo

Crate metadata from crates.io search API.

CratesIoRegistry

Client for interacting with crates.io registry.

DartDependency

DartEcosystem

DartFormatter

DartParseResult

DartVersion

EcosystemRegistry

Registry for all available ecosystems.

GemInfo

Gem metadata from rubygems.org.

GemfileLockParser

Gemfile.lock parser.

GoDependency

A dependency from a go.mod file.

GoEcosystem

Go modules ecosystem implementation.

GoParseResult

Result of parsing a go.mod file.

GoRegistry

Client for interacting with proxy.golang.org.

GoVersion

Version information from proxy.golang.org.

GradleDependency

GradleEcosystem

GradleFormatter

GradleParseResult

GradleVersion

HttpCache

HTTP cache with ETag and Last-Modified validation.

MavenCentralRegistry

MavenDependency

MavenEcosystem

MavenFormatter

MavenParseResult

MavenVersion

NpmDependency

Parsed dependency from package.json with position tracking.

NpmEcosystem

npm ecosystem implementation.

NpmPackage

Package metadata from npm registry.

NpmParseResult

Result of parsing a package.json file.

NpmRegistry

Client for interacting with the npm registry.

NpmVersion

Version information for an npm package.

PackageInfo

ParseResult

Result of parsing a Cargo.toml file.

ParsedDependency

Parsed dependency from Cargo.toml with position tracking.

PubDevRegistry

PubspecLockParser

PypiDependency

Parsed dependency from pyproject.toml with position tracking.

PypiEcosystem

PyPI ecosystem implementation.

PypiParser

Parser for Python pyproject.toml files.

PypiRegistry

Client for interacting with the PyPI registry.

PypiVersion

Version information for a package from PyPI.

RubyGemsRegistry

Client for interacting with rubygems.org registry.

Enums

DependencyGroup

Gem group classification.

DependencySection

Section in Cargo.toml where a dependency is declared.

DependencySource

Source location of a dependency.

DepsError

Core error types for deps-lsp.

GoDirective

Go module directive types.

NpmDependencySection

Section in package.json where a dependency is declared.

PypiDependencySection

Section in pyproject.toml where a dependency is declared.

Functions

parse_cargo_toml

Parses a Cargo.toml file and extracts all dependencies with positions.

parse_gemfile

Parses a Gemfile and extracts all dependencies with positions.

parse_go_mod

Parses a go.mod file and extracts all dependencies with positions.

parse_gradle

parse_package_json

Parses a package.json file and extracts all dependencies with positions.

parse_pom_xml

parse_pubspec_yaml

register_ecosystems

Registers all enabled ecosystems.

Type Aliases

Result

Convenience type alias for Result<T, DepsError>.